requirements.txt
pyrofork
TgCrypto
pymongo[srv]
aiohttp
Flask==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
werkzeug==2.0.2
itsdangerous==2.0.1
python-telegram-bot==20.7
requests